requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
sqlparse==0.4.2
